# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface TranscriptMessage {
  role: string;
  content: string;
}

export interface DepartmentScore {
  departmentId: string;
  departmentName: string;
  score: number;
  justification: string;
}

export interface BusinessCase {
  departmentId: string;
  departmentName: string;
  brief: string;
}

export interface ApplicationInput {
  companyName: string;
  website?: string;
  stage?: string;
  teamSize?: string;
  transcript: TranscriptMessage[];
}

export type ApplicationUpdateInputStatus = typeof ApplicationUpdateInputStatus[keyof typeof ApplicationUpdateInputStatus];


export const ApplicationUpdateInputStatus = {
  pending: 'pending',
  routed: 'routed',
  shortlisted: 'shortlisted',
  accepted: 'accepted',
  declined: 'declined',
  archived: 'archived',
} as const;

export interface ApplicationUpdateInput {
  status?: ApplicationUpdateInputStatus;
  notes?: string;
}

export interface ApplicationSummary {
  id: string;
  companyName: string;
  website?: string;
  stage?: string;
  teamSize?: string;
  status: string;
  createdAt: string;
  trackingToken: string;
  notes?: string;
  departmentScores?: DepartmentScore[];
}

export type ApplicationStructuredData = { [key: string]: unknown };

export interface Application {
  id: string;
  companyName: string;
  website?: string;
  stage?: string;
  teamSize?: string;
  status: string;
  createdAt: string;
  trackingToken: string;
  transcript: TranscriptMessage[];
  structuredData?: ApplicationStructuredData;
  departmentScores?: DepartmentScore[];
  businessCases?: BusinessCase[];
  notes?: string;
}

export interface ApplicationTracking {
  companyName: string;
  status: string;
  createdAt: string;
  departmentScores?: DepartmentScore[];
}

export interface ApiError {
  error: string;
}