# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: applications
    description: Startup application operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /applications:
    post:
      operationId: submitApplication
      tags: [applications]
      summary: Submit a startup application
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApplicationInput"
      responses:
        "201":
          description: Application submitted and processed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Application"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
    get:
      operationId: listApplications
      tags: [applications]
      summary: List all applications
      responses:
        "200":
          description: List of applications
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ApplicationSummary"

  /applications/{id}:
    get:
      operationId: getApplication
      tags: [applications]
      summary: Get a single application with full AI analysis
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Application detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Application"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
    patch:
      operationId: updateApplication
      tags: [applications]
      summary: Update application status and/or notes
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApplicationUpdateInput"
      responses:
        "200":
          description: Updated application
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Application"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"

  /applications/track/{token}:
    get:
      operationId: trackApplication
      tags: [applications]
      summary: Get public tracking view for an application by tracking token
      parameters:
        - name: token
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Application tracking info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApplicationTracking"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    TranscriptMessage:
      type: object
      properties:
        role:
          type: string
        content:
          type: string
      required:
        - role
        - content

    DepartmentScore:
      type: object
      properties:
        departmentId:
          type: string
        departmentName:
          type: string
        score:
          type: number
        justification:
          type: string
      required:
        - departmentId
        - departmentName
        - score
        - justification

    BusinessCase:
      type: object
      properties:
        departmentId:
          type: string
        departmentName:
          type: string
        brief:
          type: string
      required:
        - departmentId
        - departmentName
        - brief

    ApplicationInput:
      type: object
      properties:
        companyName:
          type: string
        website:
          type: string
        stage:
          type: string
        teamSize:
          type: string
        transcript:
          type: array
          items:
            $ref: "#/components/schemas/TranscriptMessage"
      required:
        - companyName
        - transcript

    ApplicationUpdateInput:
      type: object
      properties:
        status:
          type: string
          enum: [pending, routed, shortlisted, accepted, declined, archived]
        notes:
          type: string

    ApplicationSummary:
      type: object
      properties:
        id:
          type: string
        companyName:
          type: string
        website:
          type: string
        stage:
          type: string
        teamSize:
          type: string
        status:
          type: string
        createdAt:
          type: string
          format: date-time
        trackingToken:
          type: string
        notes:
          type: string
        departmentScores:
          type: array
          items:
            $ref: "#/components/schemas/DepartmentScore"
      required:
        - id
        - companyName
        - status
        - createdAt
        - trackingToken

    Application:
      type: object
      properties:
        id:
          type: string
        companyName:
          type: string
        website:
          type: string
        stage:
          type: string
        teamSize:
          type: string
        status:
          type: string
        createdAt:
          type: string
          format: date-time
        trackingToken:
          type: string
        transcript:
          type: array
          items:
            $ref: "#/components/schemas/TranscriptMessage"
        structuredData:
          type: object
          additionalProperties: true
        departmentScores:
          type: array
          items:
            $ref: "#/components/schemas/DepartmentScore"
        businessCases:
          type: array
          items:
            $ref: "#/components/schemas/BusinessCase"
        notes:
          type: string
      required:
        - id
        - companyName
        - status
        - createdAt
        - trackingToken
        - transcript

    ApplicationTracking:
      type: object
      properties:
        companyName:
          type: string
        status:
          type: string
        createdAt:
          type: string
          format: date-time
        departmentScores:
          type: array
          items:
            $ref: "#/components/schemas/DepartmentScore"
      required:
        - companyName
        - status
        - createdAt

    ApiError:
      type: object
      properties:
        error:
          type: string
      required:
        - error