# lib/api-zod/src/generated/types/application.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ApplicationStructuredData } from './applicationStructuredData';
import type { BusinessCase } from './businessCase';
import type { DepartmentScore } from './departmentScore';
import type { TranscriptMessage } from './transcriptMessage';

export interface Application {
  id: string;
  companyName: string;
  website?: string;
  stage?: string;
  teamSize?: string;
  status: string;
  createdAt: Date;
  trackingToken: string;
  transcript: TranscriptMessage[];
  structuredData?: ApplicationStructuredData;
  departmentScores?: DepartmentScore[];
  businessCases?: BusinessCase[];
  notes?: string;
}