# lib/api-zod/src/generated/types/applicationInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { TranscriptMessage } from './transcriptMessage';

export interface ApplicationInput {
  companyName: string;
  website?: string;
  stage?: string;
  teamSize?: string;
  transcript: TranscriptMessage[];
}