# lib/api-zod/src/generated/types/applicationSummary.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { DepartmentScore } from './departmentScore';

export interface ApplicationSummary {
  id: string;
  companyName: string;
  website?: string;
  stage?: string;
  teamSize?: string;
  status: string;
  createdAt: Date;
  trackingToken: string;
  notes?: string;
  departmentScores?: DepartmentScore[];
}