# lib/api-zod/src/generated/types/applicationTracking.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { DepartmentScore } from './departmentScore';

export interface ApplicationTracking {
  companyName: string;
  status: string;
  createdAt: Date;
  departmentScores?: DepartmentScore[];
}