# lib/api-zod/src/generated/types/applicationUpdateInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ApplicationUpdateInputStatus } from './applicationUpdateInputStatus';

export interface ApplicationUpdateInput {
  status?: ApplicationUpdateInputStatus;
  notes?: string;
}