# lib/api-zod/src/generated/types/businessCase.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface BusinessCase {
  departmentId: string;
  departmentName: string;
  brief: string;
}