# lib/api-zod/src/generated/types/departmentScore.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface DepartmentScore {
  departmentId: string;
  departmentName: string;
  score: number;
  justification: string;
}